pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-commons"
version = "0.1.0"
description = "A federated personal knowledge management system"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
]
requires-python = ">=3.9"
dependencies = [
    "rdflib>=6.0.0",
    "rdflib-jsonld>=0.6.2",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "requests>=2.28.0",
    "anthropic>=0.8.0",
    "langchain>=0.1.0",
    "qdrant-client>=1.6.0",
    "chromadb>=0.4.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.0",
    "markdown>=3.4.0",
    "beautifulsoup4>=4.12.0",
    "python-frontmatter>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
knowledge-commons = "knowledge_commons.cli:main"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"